Initial commit
Original code from Steve and I's Senior Design project at the University of Toledo.  The project was tasked to autonomously recognize and pinpoint a target object after having successfully navigated itself through a field of obstacles.  We won.

This code will be heavily altered for Steve and I's latest project but will serve as a good starting point.